Proposal for our own paginator for query string.
Currently the Fractal Illuminate adapter fails to take the query string
parameters required.

The idea I am using is that rather then creating a new instance of the
paginator just using the one that is already provided. The wrapper just
makes it bind to the contract of PaginatorInterface.

This currently breaks tests as I have been going round and round trying
to fix it, sorry for this.