Error out when a JsonResponse is returned.
JsonResponse is not supported due to Dingo needing to handle the formatting
of the original response. While it did work, errors would be encountered
when attempt to perform internal requests on an endpoint that returned
a JsonResponse.

Use of standard response objects is encouraged, or just return whatever it
is as it is. For more information refer to the Wiki.

Reference issue #879.