Replace the controller dispatcher for API routes.
The controller dispatcher still used the Laravel router which did not have the
appropriate middlewares defined on it. A new RequestWasMatched event has been
added that is fired when an API request is matched. We are hooking into this
event so that we can swap out the controller dispatcher with one that uses
the adapters router instance instead.