List the unroutable helpers methods in ControllerInspector.
This is a throwback to when it was an abstract controller that was to
be extended. When using the Router::controller method the Helpers
trait has several public get methods that are added as routes.

Overriding the inspector and adding in an array of methods that
are not routable is the best solution for this.

Reference issue #937.