Boot parent provider after.
Opting to boot the parent provider after prevents the router binding to be fired
too early. This was causing issues primarily if the service provider was added
to the config after all the other providers.

This caused issues primarily with tests because of the execution order of the
Kernel. Middlewares were not registered on the router instance.

Reference issue #964.