Clone router prior to dispatching Laravel request.
This is a short-term fix and needs a little more investigation into possible
performance issues with cloning the router before every dispatch.

This is probably something that could be done primarily for the dispatcher
or even re-work the internal dispatcher so that it doesn't even utilzie
the actual Laravel router but instead uses its own.

Reference issue #964 and #966.